fix(saas): add saas subcmd
